requirements.txt
pip~=20.2.3
idna~=3.8
multidict~=6.0.5
attrs~=24.2.0
aiogram~=3.12.0
pydantic~=2.8.2
aiofiles~=23.2.1
certifi~=2024.7.4
aiohttp~=3.10.5
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
aiohappyeyeballs~=2.4.0
urllib3~=2.2.2
setuptools~=49.2.1
python-dotenv~=1.0.1
requests~=2.32.3